Script checks matrix size before generating a Seurat object.
Apparently, matrices with greater than approximately 2 billion observations cannot be imported into a Seurat object.  Consequently, the matrix size is checked and, depending on the size, is either imported or not imported.  In situations when the matrix is not imported, a README.txt file is generated which explains that size limit for Seurat was exceeded.